% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extras.R
\name{extract_min6_series}
\alias{extract_min6_series}
\title{Extract single sseries from existing MIN-6 files}
\usage{
extract_min6_series(
  base_path,
  target_row = "2\\\\. ISKALEC PONOVNE ZAPOSLITVE"
)
}
\arguments{
\item{base_path}{where the files are hidden}

\item{target_row}{text expected in first column or target row}
}
\value{
dataframe with period and value columns, period being 2025M12 format
}
\description{
This currently extracts column B, but i can extend to add the column
parameter as well. The whole thing depends on the year-month folder
structure, so this is a one off for these particular files. just saving
it to the package to be more organized.
}
