% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_writing.R
\name{insert_data_points}
\alias{insert_data_points}
\title{Insert datapoints into data_point table}
\usage{
insert_data_points(meta, df, con, schema = "platform")
}
\arguments{
\item{meta}{is the dataframe in the package with the table names}

\item{df}{dataframes with the data_points
output of \link[ZRSZfetchR]{zrsz_bo_excel_parser} or similar.}

\item{con}{connection to database}

\item{schema}{is the database schema}
}
\value{
nothing, just some printing along the way
}
\description{
So, the function extracts and preps the data with \link[ZRSZfetchR]{prepare_data_table}
and writes it to a temporary table in the database.
}
\details{
It inserts any new periods into the period table,
adds the data points to the data point table.
}
