% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_writing.R
\name{insert_new_data}
\alias{insert_new_data}
\title{Insert new data for a table i.e. a vintage}
\usage{
insert_new_data(meta, df, con, schema)
}
\arguments{
\item{meta}{is the dataframe in the package with the table names}

\item{df}{dataframes with the data_points}

\item{con}{connection to database}

\item{schema}{database schema}
}
\value{
list of tables with counts for each inserted row.
}
\description{
When new data for a table (one of the Excel's) is added, these are new
vintages. This function inserts a set of new vintages and their corresponding
data points to the database. It is possible to only have new monthly not annual
vintages.
}
