% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_writing.R
\name{insert_new_source}
\alias{insert_new_source}
\title{Insert new source.}
\usage{
insert_new_source(con)
}
\arguments{
\item{con}{connection to the database.}
}
\value{
nothing
}
\description{
This is a one off, but is here for completeness. Used to insert the ZRSZ source
into the database. Also adds the cover category for the source.
}
