% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_writing.R
\name{insert_new_table_structures}
\alias{insert_new_table_structures}
\title{Insert table structure data for the tables}
\usage{
insert_new_table_structures(meta, con)
}
\arguments{
\item{meta}{is the dataframe in the package with the table names}

\item{con}{connection to the database.}
}
\value{
nothing
}
\description{
This is here for the record/reprodcibility, because it was only done once.
WHen the five tables were first added to the database, a set of nine
tables needed to be populated with appropriate metadata about those tables.
This umbrella function calls the respective SQL functions for each
of the nine tables and the table preparation functions to insert the data.
}
