% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_structure.R
\name{prepare_category_relationship_table}
\alias{prepare_category_relationship_table}
\title{Prepare table to insert into `category_relationship` table}
\usage{
prepare_category_relationship_table(meta, con)
}
\arguments{
\item{meta}{dataframe with code, name, url and note columns}

\item{con}{connection to the database}
}
\value{
a dataframe with the `id`, `parent_id`, `source_id` for each relationship
betweeh categories
}
\description{
Helper function that manually prepares the category_relationship table.
Returns table ready to insert into the `category_relationship` table with the db_writing family
of functions from SURSfetchR using \link[SURSfetchR]{sql_function_call}
}
