% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_structure.R
\name{prepare_category_table}
\alias{prepare_category_table}
\title{Prepare table to insert into `category` table}
\usage{
prepare_category_table(meta, con, schema = "platform")
}
\arguments{
\item{meta}{dataframe with code, name, url and note columns}

\item{con}{connection to the database}

\item{schema}{schema name}
}
\value{
a dataframe with the `id`, `name`, `source_id` for each category that
the table is a member of.
}
\description{
Helper function that manually prepares the category table with field ids and
their names. Returns table ready to insert into the `category` table with the db_writing family
of functions from `UMARimportR`.
}
