% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_data.R
\name{prepare_data_table}
\alias{prepare_data_table}
\title{Get and prepare data for import}
\usage{
prepare_data_table(meta, df, con)
}
\arguments{
\item{meta}{dataframe with code, name, url and note columns}

\item{df}{list with at least monthly and annual dataframes with the data_points
output of \link[MFfetchR]{mf_excel_parser}.}

\item{con}{connection to database}
}
\value{
a dataframe with the period_id, value and id values for all the vintages in the table.
}
\description{
Prepares the timeseries data for importing into the database.
}
