% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_data.R
\name{prepare_datapoint_table}
\alias{prepare_datapoint_table}
\title{Prepare datapoint table for insertion}
\usage{
prepare_datapoint_table(df, meta, con, schema)
}
\arguments{
\item{df}{data frame with the data points from \link[ZRSZfetchR]{get_datapoints}}

\item{meta}{dataframe row with the table metadata, with columns `code`}

\item{con}{connection to the database}

\item{schema}{schema name}
}
\value{
a list with data, table_id, dimension_names and dimension_ids
}
\description{
Prepare the data points for insertion into the database. This is after
the vintages have been inserted already, all of which happens in
\link[ZRSZfetchR]{ZRSZ_import_data_points}
}
