% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_structure.R
\name{prepare_series_table}
\alias{prepare_series_table}
\title{Prepare table to insert into `series` table}
\usage{
prepare_series_table(meta, con)
}
\arguments{
\item{meta}{dataframe with code, name, url and note columns}

\item{con}{connection to the database}
}
\value{
a dataframe with the following columns: `name_long`, `code`,
`unit_id`, `table_id` and `interval_id`for each series in the table
well as the same number of rows as there are series
}
\description{
This currently only works for a single series table, since that is the only
case i have right now, but will expand if necessary.
}
