% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_structure.R
\name{prepare_table_dimensions_table}
\alias{prepare_table_dimensions_table}
\title{Prepare table to insert into `table_dimensions` table}
\usage{
prepare_table_dimensions_table(meta, con)
}
\arguments{
\item{meta}{dataframe with code, name, url and note columns}

\item{con}{connection to the database}
}
\value{
a dataframe with the `table_id`, `dimension_name`, `time` columns for
each dimension of this table.
}
\description{
Helper function that manually prepares the table_dimensions table.
Returns table ready to insert into the `table_dimensions`table with the
db_writing family of functions.
}
