% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_data.R
\name{prepare_vintage_table}
\alias{prepare_vintage_table}
\title{Prepare table to insert into `vintage` table}
\usage{
prepare_vintage_table(meta, df, con, schema)
}
\arguments{
\item{meta}{dataframe with code, name, url and note columns}

\item{df}{parsed data (with at least a period column)}

\item{con}{connection to database}

\item{schema}{schema name}
}
\value{
vintage table ready to insert
}
\description{
Helper function that prepares the vintage table with the new vintages.
First checks if new vintages are even necessary, getting
to get the most recent month from the parsed data, and then checking with the most
recent period in the database. If there is new data,
Prepare a dataframe with the `series_id`s of the table and the current time
as their publication date - since we don't have anything better.
}
\details{
Returns table ready to insert into the `vintage`table with the
db_writing family of functions.
}
