% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extras.R
\name{update_min6_series}
\alias{update_min6_series}
\title{Upsert data from MIN6 to umar data}
\usage{
update_min6_series(con)
}
\arguments{
\item{con}{database connection}
}
\value{
invisibly the updated table
}
\description{
checks most recent period in the database and then if there is new data
available in the data folder. if su updates the umar-data table.
}
