% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_data.R
\name{vintage_table}
\alias{vintage_table}
\title{Prepare vintage table}
\usage{
vintage_table(tbl_id, con)
}
\arguments{
\item{tbl_id}{numeric table id}

\item{con}{connection to the database.}

\item{meta}{dataframe with code, name, url and note columns}
}
\value{
data frame with `series_id` and `published` columns
}
\description{
Helper function preparing for the vintage table for a specific table.
Uses current time as `published`
}
\keyword{internal}
