% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excelling.R
\name{zrsz_bo_excel_parser}
\alias{zrsz_bo_excel_parser}
\title{ZRSZ BO Execl table parser}
\usage{
zrsz_bo_excel_parser(file_path)
}
\arguments{
\item{file_path}{path to the excel file}
}
\value{
dataframe with period and value ready to import into database
}
\description{
Parsing function to extract the data in the Gibanje registrirane brezposelnosti
po mesecih Excel file, usually published [here](https://www.ess.gov.si/partnerji/trg-dela/trg-dela-v-stevilkah/registrirana-brezposelnost/),
The parser joins the tables from separate sheets and should work indefinitely
even if more sheets are added as long as the strucutre of the tables remains the
same. The data is cleaned of missing values and perpared in long format ready for
importing into the database.
}
\details{
Hardcoded failure points:
* The strucutre of the table
* Skipping a single line at the top
}
