% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excelling.R
\name{zrsz_bo_os_excel_parser}
\alias{zrsz_bo_os_excel_parser}
\title{ZRSZ BO OS Excel table parser}
\usage{
zrsz_bo_os_excel_parser(file_path)
}
\arguments{
\item{file_path}{path to the excel file}
}
\value{
dataframe with period and value ready to import into database
}
\description{
Parsing function to extract the data in the Stopnja registrirane brezposelnosti
po občinah in statističnih regijah Excel file. The parser extracts data for
Slovenia from each sheet and combines them into a single time series.
}
