% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scripting.R
\name{zrsz_bo_script}
\alias{zrsz_bo_script}
\title{Script for the BO table}
\usage{
zrsz_bo_script(con)
}
\arguments{
\item{con}{connection to database}
}
\value{
number of rows inserted
}
\description{
this is an automation script that does the following: downloads the xlsx file
on the hardcoded which changes with the year. (in January, the tryCatch construct
should use the alternative url instead, to get the previous year).
The file is downloaded to a tempfile, then the data is parsed and written into
the database.
}
