% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excelling.R
\name{zrsz_bo_trajanje_excel_parser}
\alias{zrsz_bo_trajanje_excel_parser}
\title{ZRSZ BO trajanje Excel table parser}
\usage{
zrsz_bo_trajanje_excel_parser(file_path)

zrsz_bo_trajanje_excel_parser(file_path)
}
\arguments{
\item{file_path}{path to the excel file}
}
\value{
dataframe with period and value ready to import into database

dataframe with period and value ready to import into database
}
\description{
Parsing function to extract the data in the Stopnja registrirane brezposelnosti
po trajanju Excel file. The parser extracts data for
all four education levels from each sheet (from 2014 onwards)
and combines them into a single time series. Well, four actually.

Parsing function to extract the data in the Stopnja registrirane brezposelnosti
po trajanju Excel file. The parser extracts data for
all four education levels from each sheet (from 2014 onwards)
and combines them into a single time series. Well, four actually.
}
