% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excelling.R
\name{zrsz_dd_excel_parser}
\alias{zrsz_dd_excel_parser}
\title{ZRSZ DD Excel table parser}
\usage{
zrsz_dd_excel_parser(file_path)
}
\arguments{
\item{file_path}{path to the excel file}
}
\value{
dataframe with period and value ready to import into database
}
\description{
Parsing function to extract the data in the delovna dovoljenja (po spolu)
 Excel file. The parser extracts data for the SKupaj series
from each sheet and combines them into one long time series.
We are ignoring the sheets before 2012, because they don't have
data for all 12 months for some reason...
}
