% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importing.R
\name{ZRSZ_import_data_points}
\alias{ZRSZ_import_data_points}
\title{Insert data points from ZRSZ}
\usage{
ZRSZ_import_data_points(meta, con, schema = "platform")
}
\arguments{
\item{meta}{ZRSZ row with metadata for the table, see raw_data}

\item{con}{Database connection}

\item{schema}{Schema name}
}
\value{
Insertion results (invisibly)
}
\description{
Function to prepare and insert ZRSZ data points. The function first prepares
the required vintages and inserts them, then prepares the data points
table and isnerts it. The function returns the results invisibly.
}
\details{
This is a ZRSZ specific function, which should be followed by the generic
UMARimportR function to write the vintage hashes and clean up redundant
vintages.
}
