% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importing.R
\name{ZRSZ_import_structure}
\alias{ZRSZ_import_structure}
\title{Import structural metadata for a ZRSZ table}
\usage{
ZRSZ_import_structure(meta, con, schema = "platform", keep_vintage = FALSE)
}
\arguments{
\item{meta}{row with metadata for the table, see raw_data}

\item{con}{connection to database}

\item{schema}{schema name, defaults to platform}

\item{keep_vintage}{logical indicating whether to keep vintages, defaults to F}
}
\value{
nothing
}
\description{
Umbrella function that prepares and import all the metadata tables into
the database.
}
