% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excelling.R
\name{zrsz_tok_excel_parser}
\alias{zrsz_tok_excel_parser}
\title{ZRSZ TOK Excel table parser}
\usage{
zrsz_tok_excel_parser(file_path)
}
\arguments{
\item{file_path}{path to the excel file}
}
\value{
dataframe with period and value ready to import into database
}
\description{
Parsing function to extract the data in the prijave odjave
 Excel file. The parser extracts data for eleven series
from each sheet and combines them into long time series
}
