% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_structure.R
\name{prepare_category_table_table}
\alias{prepare_category_table_table}
\title{Prepare table to insert into `category_table` table}
\usage{
prepare_category_table_table(meta, con, schema = "platform")
}
\arguments{
\item{meta}{dataframe with code, name, url and note columns}

\item{con}{connection to the database}

\item{schema}{schema name}
}
\value{
a dataframe with the `category_id` `table_id` and `source_id` columns for
each table-category relationship.
}
\description{
Helper function that manually prepares the category_table table.
Returns table ready to insert into the `category_table` table with the db_writing family
of functions from `UMARimportR`.
A single table can have multiple parents - meaning
it is member of several categories (usually no more than two tho). .
}
