% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_structure.R
\name{prepare_table_table}
\alias{prepare_table_table}
\title{Prepare table to insert into `table` table}
\usage{
prepare_table_table(meta, con)
}
\arguments{
\item{meta}{dataframe with code, name, url and note columns}

\item{con}{connection to the database}
}
\value{
a dataframe with the `code`, `name`, `source_id`, `url`, and `notes` columns
for this table.
}
\description{
Helper function that manually prepares the table table.
Returns table ready to insert into the `table` table with the db_writing family
of functions from SURSfetchR using \link[SURSfetchR]{sql_function_call}
}
